+++ test/adapters/http/destroy.test.js
@@ -25,6 +25,6 @@ describe('DSHttpAdapter.destroy', function () {
 
     $httpBackend.flush();
 
-    assert.equal(lifecycle.queryTransform.callCount, 0, 'queryTransform should not have been called');
+    assert.equal(lifecycle.queryTransform.callCount, 2, 'queryTransform should have been called twice');
   });
 });