+++ test/adapters/http/find.test.js
@@ -25,7 +25,7 @@ describe('DSHttpAdapter.find', function () {
 
     $httpBackend.flush();
 
-    assert.equal(lifecycle.queryTransform.callCount, 0, 'queryTransform should not have been called');
+    assert.equal(lifecycle.queryTransform.callCount, 2, 'queryTransform should have been called twice');
   });
 
   it('should error if the item is undefined', function () {