+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+##### 2.2.2 - 04 March 2015
+
+###### Backwards compatible bug fixes
+- #312 - UMDifying js-data-angular  
+- #313 - DSHttpAdapter#find/create/update/destroy do not call queryTransform
+
 ##### 2.2.1 - 25 February 2015
 
 ###### Backwards compatible bug fixes